Make basic authenticator default
Jwt authenticator is undocumented and not expected to be used. Using it
as the default authenticator doesn't make sense.
This change uses the basic authenticator (added since #1098) as the
default authenticator.

If no configuration is set, behavior is kept same (no authentications
happen).

Current BasicAuthenticator does not support multi-tenancy (multiple
site_id) but it can be added in the future.

This change includes documentation and test changes so that system
plugin mechanism works intentionally for AuthenticatorFactory SPI.

This change is expected to be merged for v0.10 which deprecates admin
API.